% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create_simple_LMDI}
\alias{create_simple_LMDI}
\title{Create a simple LMDI data frame}
\usage{
create_simple_LMDI()
}
\value{
a data frame of example LMDI data
}
\description{
This function creates a simple LMDI data frame that can be used for testing or examples.
It contains \code{X} matrices for two fictitious countries (\code{AB} and \code{XY})
and four years (\code{1971-1974}).
Countries \code{AB} and \code{YZ} are identical for the purposes of illustration.
}
\examples{
library(matsindf)
DF <- create_simple_LMDI()
DF
DF$X[[1]]
}
