% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdi.R
\name{lmdi}
\alias{lmdi}
\title{Log-Mean Divisia Index (LMDI) decomposition analysis}
\usage{
lmdi(
  .lmdidata,
  time = "Year",
  X = "X",
  fillrow = NULL,
  V = "V",
  Z = "Z",
  deltaV = "dV",
  D = "D"
)
}
\arguments{
\item{.lmdidata}{a grouped data frame.
Group by columns of variables within which you want an LMDI analysis conducted.
\code{time} should not be one of the grouping variables.}

\item{time}{the name of the column in \code{.lmdidata} that contains times at which
data are available (a string).
Default is "\code{Year}".}

\item{X}{the name (as a string) of a column in \code{.lmdidata} containing
\code{X} matrices with
named rows representing subcategories of the energy aggregate (\code{V}) and
named columns representing factors contributing to changes in \code{V} over time.
Default is "\code{X}".}

\item{fillrow}{a row vector of type \code{matrix} passed to \code{\link{Z_byname}}.
(See \code{\link{Z_byname}} for details.)}

\item{V}{the name for the \code{V} output column (a string).
Default is "\code{V}".}

\item{Z}{the name for the \code{Z} output column (a string).
Default is "\code{Z}".}

\item{deltaV}{the name for the \code{deltaV} output column (a string).
Default is "\code{dV}".}

\item{D}{the name for the \code{D} output column (a string).
Default is "\code{D}".}
}
\value{
a data frame containing several columns.
}
\description{
Performs log-mean divisia index decomposition analysis on a suitably-formatted data frame.
}
