\name{LMDIR-package}
\alias{LMDIR-package}
\alias{LMDIR}
\docType{package}
\title{
\packageTitle{LMDIR}
}
\description{
\packageDescription{LMDIR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LMDIR}
\packageIndices{LMDIR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LMDIR}

Maintainer: \packageMaintainer{LMDIR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
