% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Z_byname}
\alias{Z_byname}
\title{Calculates a Z matrix from \code{X_0} and \code{X_T} matrices}
\usage{
Z_byname(X_0, X_T)
}
\arguments{
\item{X_0}{an \code{X} matrix for time \code{0}}

\item{X_T}{an \code{X} matrix for time \code{T}}
}
\value{
a \code{Z} matrix
}
\description{
This function fully accounts for all degenerate cases
found in Table 2, p. 492 of
B.W. Ang and F.Q. Zhang and Ki-Hong Choi, 1998,
Factorizing changes in energy and environmental indicators through decomposition,
Energy, Volume 23, Number 6, pp. 489-495.
}
