% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create0Tcolumns}
\alias{create0Tcolumns}
\title{Make columns for time 0 and time T}
\usage{
create0Tcolumns(XvV, time_colname, X_colname = "X", v_colname = "v",
  V_colname = "V", zero_suffix = "_0", T_suffix = "_T")
}
\arguments{
\item{XvV}{a data frame containing \code{X}, \code{v}, and \code{V} columns.}

\item{time_colname}{the name of the time column (a string)}

\item{X_colname}{the name of the \code{X} column (default is "\code{X}")}

\item{v_colname}{the name of the \code{v} column (default is "\code{v}")}

\item{V_colname}{the name of the \code{V} column (default is "\code{V}")}

\item{zero_suffix}{suffix for "\code{0}" variables (default is "\code{_0}")}

\item{T_suffix}{suffix for "\code{T}" variables (default is "\code{_T}")}
}
\value{
a data frame containing metadata (\code{time_colname} and grouping variables)
        and columns for X_0, v_0, V_0, X_T, v_T, and V_T.
}
\description{
Make columns for time 0 and time T
}
