% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{create_simple_LMDI}
\alias{create_simple_LMDI}
\title{Create a simple LMDI data frame}
\usage{
create_simple_LMDI()
}
\value{

}
\description{
This function creates a simple LMDI data frame that can be used for testing or examples.
}
