% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdi.R
\name{lmdi}
\alias{lmdi}
\title{Log-Mean Divisia Index (LMDI) decomposition analysis}
\usage{
lmdi(
  .lmdidata,
  time = "Year",
  X = "X",
  fillrow = NULL,
  weights = c("LMDI-I", "LMDI-II"),
  V = "V",
  Z = "Z",
  deltaV = "dV",
  D = "D"
)
}
\arguments{
\item{.lmdidata}{a grouped data frame.
Group by columns of variables within which you want an LMDI analysis conducted.
\code{time} should not be one of the grouping variables.}

\item{time}{the name of the column in \code{.lmdidata} that contains times at which
data are available (a string).
Default is "\code{Year}".}

\item{X}{the name (as a string) of a column in \code{.lmdidata} containing
\code{X} matrices with
named rows representing subcategories of the energy aggregate (\code{V}) and
named columns representing factors contributing to changes in \code{V} over time.
Default is "\code{X}".}

\item{fillrow}{a row vector of type \code{matrix} passed to \code{\link[=Z_byname]{Z_byname()}}.
(See \code{\link[=Z_byname]{Z_byname()}} for details.)}

\item{weights}{the name of the weights type chosen for the LMDI calculations.
\code{weights} should be one of "LMDI-I" or "LMDI-II". Default is "LMDI-I".}

\item{V}{the name for the \code{V} output column (a string).
Default is "\code{V}".}

\item{Z}{the name for the \code{Z} output column (a string).
Default is "\code{Z}".}

\item{deltaV}{the name for the \code{deltaV} output column (a string).
Default is "\code{dV}".}

\item{D}{the name for the \code{D} output column (a string).
Default is "\code{D}".}
}
\value{
a data frame containing several columns, including:
* \code{Z}: The decomposed Z matrix.
* \code{dV}: The additive change in V calculated from Z.
* \code{D}: The multiplicative changes in V calculated from Z.
* \code{dV_agg} and \code{D_agg}; The aggregate changes.
* \code{dV_cum} and \code{D_cum}: The cumulative sums and products.
}
\description{
Performs log-mean divisia index decomposition analysis on a suitably-formatted data frame.
The results are provided for two weight types, LMDI-I and LMDI-II, suffixed with \verb{_I} and \verb{_II}.
}
