% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdi.R
\name{lmdi}
\alias{lmdi}
\title{Log-Mean Divisia Index (LMDI) decomposition analysis}
\usage{
lmdi(.lmdidata, time_colname = "Year", X_colname = "X", pad = c("tail",
  "head"), pad.value = NA, deltaV_colname = "dV", D_colname = "D")
}
\arguments{
\item{.lmdidata}{a grouped data frame.
Group by columns of variables within which you want an LMDI analysis conducted.
\code{time_colname} should not be one of the grouping variables.}

\item{time_colname}{the name of the column in \code{.data} that contains times at which
data are available (a string).
Default is "\code{Year}".}

\item{X_colname}{the name (as a string) of a column in \code{.DF} containing
\strong{\code{X}} matrices with
named rows representing subcategories of the energy aggregate (\code{V}) and
named columns representing factors contributing to changes in \code{V} over time.}

\item{pad}{either "\code{tail}" or "\code{head}" to indicate whether the first or last row,
respectively, should contain \code{pad.value}.}

\item{pad.value}{the value to be used in the padding row.  Default is \code{NA}.}

\item{deltaV_colname}{the name for the \code{deltaV} column (a string).}

\item{D_colname}{the name for the \code{D} column (a string).}
}
\value{
a data frame containing several columns.
}
\description{
Performs log-mean divisia index decomposition analysis on a suitably-formatted data frame.
}
