\name{LMDIR-package}
\alias{LMDIR-package}
\alias{LMDIR}
\docType{package}
\title{
\packageTitle{LMDIR}
}
\description{
\packageDescription{LMDIR}
}
\details{
    This package implements log-mean divisia index decomposition analysis in R (LMDIR).
    The calculations are based on a 2005 paper by B.W. Ang.
}
\author{
\packageAuthor{LMDIR}

\strong{Maintainer}: \packageMaintainer{LMDIR}
}
\references{
B.W. Ang. The LMDI approach to decomposition analysis: A practical guide.
    Energy Policy, 33:867-871, 2005.
}
\keyword{ algebra, array, manip }
\examples{
# Add examples here
}