% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XGH}
\alias{XGH}
\title{X matrices for Ghana}
\format{
A data frame with 43 rows and 3 variables:
\describe{
  \item{Country}{The country, Ghana}
  \item{Year}{The year for this row of the data frame}
  \item{X}{The \code{X} matrix for the year of this row of the data frame}
}
}
\source{
\url{http://doi.org/xxxxxx}
}
\usage{
XGH
}
\description{
A dataset containing \code{X} matrices for Ghana over a variety of years.
}
\details{
These data were used to create Figure 7 and Table 6 in the paper

M.K. Heun and P.E. Brockway. 2019. "Meeting 2030 primary energy and economic growth goals: Mission impossible?".
Applied Energy, Vol. xxxx, No. xxxx, pp. xxxx-xxxx.
}
\keyword{datasets}
