% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Z_byname}
\alias{Z_byname}
\title{Z matrix from \code{X_0} and \code{X_T} matrices}
\usage{
Z_byname(X_0, X_T)
}
\arguments{
\item{X_0}{an \code{X} matrix for initial time \code{0}}

\item{X_T}{an \code{X} matrix for final time \code{T}}
}
\value{
a \code{Z} matrix
}
\description{
The formula for Z is \deqn{Z_ij = logmean(v_Ti1, v_0i1) * log(X_Tij / X_0ij)}
where \code{v} is a column vector formed by row products of \code{X}.
The \code{T} and \code{0} subscripts on \code{X} and \code{v} indicate an initial time (\code{0})
and a final time (\code{T}).
}
\details{
The nomenclature for this function comes from
\href{https://doi.org/10.1016/s0360-5442(98)00016-4}{Ang, Zhang, and Choi (1998)}.
This function fully accounts for the degenerate cases
found in Table 2, p. 492 of
\href{https://doi.org/10.1016/s0360-5442(98)00016-4}{Ang, Zhang, and Choi (1998)}.
We're employing the method suggested by
\href{https://doi.org/10.1016/j.enpol.2004.11.010}{Wood and Lenzen (2006)}.
}
